requirements.txt
boost==0.1
cmake==0.7.1
boost-py
face_recognition_models
face_recognition
dlib>=19.3.0
dropbox
numpy
pytz
timezonefinderl
pillow
piexif
geopy
python-resize-image
flask
Flask-Limiter
gunicorn
redis
rq
PyYAML
